moving code to src file
