+++ src/stripRoutine.js
